refactor register user
